requirements.txt
Flask==1.0.2
Flask-SocketIO==3.3.2
gunicorn==19.9.0
numpy==1.16.3
Pillow==6.0.0